Update poweroff.c
root check

+++ poweroff.c
@@ -3,6 +3,11 @@
 #include <unistd.h>
 
 int main() {
+    if (geteuid() != 0) {
+        fprintf(stderr, "iinitt: please run this program as root\n");
+        return 1;
+    }
+    
     kill(1, SIGTERM);
     return 0;
 }
